Rename package names according to Eclipse naming conventions
