+++ jakarta.eclipse/org.eclipse.lsp4jakarta.lsp4e.core/META-INF/MANIFEST.MF
@@ -4,7 +4,7 @@ Bundle-Name: Eclipse LSP4Jakarta Plugin (Incubation)
 Bundle-Vendor: Eclipse LSP4Jakarta
 Bundle-SymbolicName: org.eclipse.lsp4jakarta.lsp4e.core;singleton:=true
 Bundle-Version: 0.0.1.qualifier
-Bundle-Activator: org.jakarta.lsp4e.Activator
+Bundle-Activator: org.eclipse.lsp4jakarta.lsp4e.Activator
 Require-Bundle: 
  org.eclipse.lsp4e,
  org.eclipse.core.runtime,
