+++ jakarta.eclipse/org.eclipse.lsp4jakarta.lsp4e.core/src/org/eclipse/lsp4jakarta/lsp4e/JakartaLanguageClient.java
@@ -11,7 +11,7 @@
 *     IBM Corporation - initial API and implementation
 *******************************************************************************/
 
-package org.jakarta.lsp4e;
+package org.eclipse.lsp4jakarta.lsp4e;
 
 import java.util.ArrayList;
 import java.util.List;
