--- jakarta.eclipse/org.eclipse.lsp4jakarta.lsp4e.core/src/org/jakarta/lsp4e/Activator.java
@@ -1,73 +1,0 @@
-/*******************************************************************************
-* Copyright (c) 2020 IBM Corporation and others.
-*
-* This program and the accompanying materials are made available under the
-* terms of the Eclipse Public License v. 2.0 which is available at
-* http://www.eclipse.org/legal/epl-2.0, or the Apache License, Version 2.0
-* which is available at https://www.apache.org/licenses/LICENSE-2.0.
-*
-* SPDX-License-Identifier: EPL-2.0 OR Apache-2.0
-*
-* Contributors:
-*     IBM Corporation - initial API and implementation
-*******************************************************************************/
-
-package org.jakarta.lsp4e;
-
-import org.eclipse.core.runtime.IStatus;
-import org.eclipse.core.runtime.Status;
-import org.eclipse.ui.plugin.AbstractUIPlugin;
-import org.osgi.framework.BundleContext;
-
-/**
- * The activator class controls the plug-in life cycle
- */
-public class Activator extends AbstractUIPlugin {
-
-    // The plug-in ID
-    public static final String PLUGIN_ID = "org.eclipse.lsp4jakarta.lsp4e.core"; //$NON-NLS-1$
-
-    // The shared instance
-    private static Activator plugin;
- 
-    public boolean started;
-
-    /**
-     * The constructor
-     */
-    public Activator() {
-    }
-
-    @Override
-    public void start(BundleContext context) throws Exception {
-        super.start(context);
-        plugin = this;
-        getDefault().getLog().log(new Status(IStatus.INFO, PLUGIN_ID, "Starting activator class for lsp4e.core"));
-        started = true;
-    }
-
-    @Override
-    public void stop(BundleContext context) throws Exception {
-        plugin = null;
-        started = false;
-        super.stop(context);
-    }
-
-    /**
-     * Returns the shared instance
-     *
-     * @return the shared instance
-     */
-    public static Activator getDefault() {
-        return plugin;
-    }
-
-    public static void log(IStatus status) {
-        getDefault().getLog().log(status);
-    }
-
-    public static void logException(String errMsg, Throwable ex) {
-        getDefault().getLog().log(new Status(IStatus.ERROR, PLUGIN_ID, errMsg, ex));
-    }
-
-}
